Code for PA 6 tested and working

+++ Prob-3/Prob-3.py
@@ -6,7 +6,7 @@
 
 def main():
     sum = 0
-    num = int(input('enter a positive number (negative to end):'))
+    num = float(input('enter a positive number (negative to end):'))
 
     # do not change the while loop definition below
     while True:
@@ -14,7 +14,7 @@ def main():
         if num <= 0:
               break
         sum = sum + num
-        num = int(input('enter a positive number (negative to end):'))
+        num = float(input('enter a positive number (negative to end):'))
 
     print('the sum of the numbers is:', sum)
